Split controller and driver
