+++ src/serial.rs
@@ -1,325 +1,126 @@
-use crate::param::RemainingBytes;
-use crate::param::Status;
+use crate::data::AclPacketHeader;
+use crate::driver::HciDriver;
+use crate::event::EventPacketHeader;
 use crate::FromHciBytes;
-use crate::ReadHci as _;
-use crate::WriteHci as _;
-use crate::{
-    cmd::{self},
-    data,
-    event::{CommandComplete, Event},
-    param::{self},
-    CmdError, Controller, ControllerCmdAsync, ControllerCmdSync, ControllerToHostPacket, FixedSizeValue, WithIndicator,
-};
-use core::ops::DerefMut;
-use core::{cell::RefCell, future::Future};
-use embassy_sync::{
-    blocking_mutex::raw::{NoopRawMutex, RawMutex},
-    mutex::Mutex,
-    signal::Signal,
-};
-use embedded_io::Error;
+use crate::FromHciBytesError;
+use crate::PacketKind;
+use crate::ReadHciError;
+use core::future::Future;
+use embassy_sync::{blocking_mutex::raw::RawMutex, mutex::Mutex};
+use embedded_io::ReadExactError;
 
-/// The controller state holds a number of command slots that can be used
-/// to issue commands and await responses from an underlying controller.
-///
-/// The contract is that before sending a command, a slot is reserved, which
-/// returns a signal handle that can be used to await a response.
-pub struct ControllerState<const SLOTS: usize> {
-    slots: RefCell<[CommandSlot; SLOTS]>,
-    signals: [Signal<NoopRawMutex, CommandResponse>; SLOTS],
-}
-
-pub struct CommandResponse {
-    num_hci_cmd_pkts: u8,
-    status: Status,
-    len: usize,
-}
-
-pub enum CommandSlot {
-    Empty,
-    Pending { opcode: u16, event: *mut u8 },
-}
-
-impl<const SLOTS: usize> ControllerState<SLOTS> {
-    const EMPTY_SLOT: CommandSlot = CommandSlot::Empty;
-    const EMPTY_SIGNAL: Signal<NoopRawMutex, CommandResponse> = Signal::new();
-
-    pub fn new() -> Self {
-        Self {
-            slots: RefCell::new([Self::EMPTY_SLOT; SLOTS]),
-            signals: [Self::EMPTY_SIGNAL; SLOTS],
-        }
-    }
-
-    pub fn complete(&self, evt: &CommandComplete<'_>) {
-        let mut slots = self.slots.borrow_mut();
-        for (idx, slot) in slots.iter_mut().enumerate() {
-            match slot {
-                CommandSlot::Pending { opcode, event } if *opcode == evt.cmd_opcode.to_raw() => {
-                    let data = evt.return_param_bytes.as_ref();
-
-                    // Safety: since the slot is in pending, the caller stack will be valid.
-                    unsafe { event.copy_from(data.as_ptr(), data.len()) };
-                    self.signals[idx].signal(CommandResponse {
-                        num_hci_cmd_pkts: evt.num_hci_cmd_pkts,
-                        status: evt.status,
-                        len: evt.return_param_bytes.len(),
-                    });
-                    break;
-                }
-                _ => {}
-            }
-        }
-    }
-
-    fn release_slot(&self, idx: usize) {
-        let mut slots = self.slots.borrow_mut();
-        slots[idx] = CommandSlot::Empty;
-    }
-
-    fn acquire_slot(&self, op: cmd::Opcode, event: *mut u8) -> Option<(&Signal<NoopRawMutex, CommandResponse>, usize)> {
-        let mut slots = self.slots.borrow_mut();
-        // Make sure there are no existing command with this opcode
-        for slot in slots.iter() {
-            match slot {
-                CommandSlot::Pending { opcode, event: _ } if *opcode == op.to_raw() => {
-                    return None;
-                }
-                _ => {}
-            }
-        }
-        // Reserve our slot
-        for (idx, slot) in slots.iter_mut().enumerate() {
-            match slot {
-                CommandSlot::Empty => {
-                    *slot = CommandSlot::Pending {
-                        opcode: op.to_raw(),
-                        event,
-                    };
-                    self.signals[idx].reset();
-                    return Some((&self.signals[idx], idx));
-                }
-                _ => {}
-            }
-        }
-        None
-    }
-}
-
-pub struct SerialController<M, R, W, const SLOTS: usize>
-where
-    M: RawMutex,
-    R: embedded_io_async::Read,
-    W: embedded_io_async::Write,
-{
+/// A HCI driver implementation for a split serial
+pub struct SerialHciDriver<M: RawMutex, R, W> {
     reader: Mutex<M, R>,
     writer: Mutex<M, W>,
-    slots: ControllerState<SLOTS>,
 }
 
-impl<M, R, W, const SLOTS: usize> SerialController<M, R, W, SLOTS>
-where
-    M: RawMutex,
-    R: embedded_io_async::Read,
-    W: embedded_io_async::Write,
-{
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
+pub enum WriteHciError<E: embedded_io::Error> {
+    Write(E),
+}
+
+impl<E: embedded_io::Error> embedded_io::Error for WriteHciError<E> {
+    fn kind(&self) -> embedded_io::ErrorKind {
+        match self {
+            Self::Write(e) => e.kind(),
+        }
+    }
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
+pub enum Error<E: embedded_io::Error> {
+    Read(ReadHciError<E>),
+    Write(WriteHciError<E>),
+}
+
+impl<E: embedded_io::Error> embedded_io::Error for Error<E> {
+    fn kind(&self) -> embedded_io::ErrorKind {
+        match self {
+            Self::Read(e) => e.kind(),
+            Self::Write(e) => e.kind(),
+        }
+    }
+}
+
+impl<E: embedded_io::Error> From<WriteHciError<E>> for Error<E> {
+    fn from(e: WriteHciError<E>) -> Self {
+        Self::Write(e)
+    }
+}
+
+impl<E: embedded_io::Error> From<ReadHciError<E>> for Error<E> {
+    fn from(e: ReadHciError<E>) -> Self {
+        Self::Read(e)
+    }
+}
+
+impl<E: embedded_io::Error> From<ReadExactError<E>> for Error<E> {
+    fn from(e: ReadExactError<E>) -> Self {
+        Self::Read(e.into())
+    }
+}
+
+impl<E: embedded_io::Error> From<FromHciBytesError> for Error<E> {
+    fn from(e: FromHciBytesError) -> Self {
+        Self::Read(e.into())
+    }
+}
+
+impl<M: RawMutex, R: embedded_io_async::Read, W: embedded_io_async::Write> SerialHciDriver<M, R, W> {
     pub fn new(reader: R, writer: W) -> Self {
         Self {
-            slots: ControllerState::new(),
             reader: Mutex::new(reader),
             writer: Mutex::new(writer),
         }
     }
 }
 
-impl<M, R, W, const SLOTS: usize> Controller for SerialController<M, R, W, SLOTS>
-where
-    M: RawMutex,
-    R: embedded_io_async::Read,
-    W: embedded_io_async::Write,
+impl<
+        M: RawMutex,
+        R: embedded_io_async::Read<Error = E>,
+        W: embedded_io_async::Write<Error = E>,
+        E: embedded_io::Error,
+    > HciDriver for SerialHciDriver<M, R, W>
 {
-    type Error = embedded_io::ErrorKind;
-    fn write_acl_data(&self, packet: &data::AclPacket) -> impl Future<Output = Result<(), Self::Error>> {
+    type Error = Error<E>;
+    fn read(&self, rx: &mut [u8]) -> impl Future<Output = Result<usize, Self::Error>> {
         async {
-            let mut io = self.writer.lock().await;
-            WithIndicator::new(packet)
-                .write_hci_async(io.deref_mut())
-                .await
-                .map_err(|e| e.kind())?;
-            Ok(())
-        }
-    }
+            let mut r = self.reader.lock().await;
 
-    fn write_sync_data(&self, packet: &data::SyncPacket) -> impl Future<Output = Result<(), Self::Error>> {
-        async {
-            let mut io = self.writer.lock().await;
-            WithIndicator::new(packet)
-                .write_hci_async(io.deref_mut())
-                .await
-                .map_err(|e| e.kind())?;
-            Ok(())
-        }
-    }
+            r.read_exact(&mut rx[0..1]).await?;
 
-    fn write_iso_data(&self, packet: &data::IsoPacket) -> impl Future<Output = Result<(), Self::Error>> {
-        async {
-            let mut io = self.writer.lock().await;
-            WithIndicator::new(packet)
-                .write_hci_async(io.deref_mut())
-                .await
-                .map_err(|e| e.kind())?;
-            Ok(())
-        }
-    }
-
-    fn read<'a>(&self, buf: &'a mut [u8]) -> impl Future<Output = Result<ControllerToHostPacket<'a>, Self::Error>> {
-        async {
-            loop {
-                {
-                    // Safety: we will not hold references across loop iterations.
-                    let buf = unsafe { core::slice::from_raw_parts_mut(buf.as_mut_ptr(), buf.len()) };
-                    let mut io = self.reader.lock().await;
-                    let value: ControllerToHostPacket<'a> = ControllerToHostPacket::read_hci_async(io.deref_mut(), buf)
-                        .await
-                        .map_err(|e| {
-                            #[cfg(not(feature = "defmt"))]
-                            warn!("Error reading from controller: {:?}", e);
-                            #[cfg(feature = "defmt")]
-                            warn!("Error reading from controller: {:?}", defmt::Debug2Format(&e));
-                            embedded_io::ErrorKind::Other
-                        })?;
-
-                    match value {
-                        ControllerToHostPacket::Event(ref event) => match &event {
-                            Event::CommandComplete(e) => {
-                                self.slots.complete(e);
-                                continue;
-                            }
-                            _ => return Ok(value),
-                        },
-                        _ => return Ok(value),
-                    }
+            match PacketKind::from_hci_bytes(&rx[0..1])?.0 {
+                PacketKind::Cmd => Err(Error::Read(ReadHciError::InvalidValue)),
+                PacketKind::AclData => {
+                    r.read_exact(&mut rx[1..5]).await?;
+                    let header = AclPacketHeader::from_hci_bytes_complete(&rx[1..5])?;
+                    let data_len = header.data_len();
+                    r.read_exact(&mut rx[5..5 + data_len]).await?;
+                    Ok(5 + data_len)
+                }
+                PacketKind::SyncData => unimplemented!(),
+                PacketKind::IsoData => unimplemented!(),
+                PacketKind::Event => {
+                    r.read_exact(&mut rx[1..3]).await?;
+                    let header = EventPacketHeader::from_hci_bytes_complete(&rx[1..3])?;
+                    let params_len = usize::from(header.params_len);
+                    r.read_exact(&mut rx[3..3 + params_len]).await?;
+                    Ok(3 + params_len)
                 }
             }
         }
     }
-}
 
-impl<M, R, W, C, const SLOTS: usize> ControllerCmdSync<C> for SerialController<M, R, W, SLOTS>
-where
-    M: RawMutex,
-    R: embedded_io_async::Read,
-    W: embedded_io_async::Write,
-    C: cmd::SyncCmd,
-    C::Return: FixedSizeValue,
-{
-    fn exec(&self, cmd: &C) -> impl Future<Output = Result<C::Return, CmdError<Self::Error>>> {
+    fn write(&self, tx: &[u8]) -> impl Future<Output = Result<(), Self::Error>> {
         async {
-            // TODO: Something more appropriately sized matching the expected return lenght
-            let mut retval: [u8; 255] = [0u8; 255];
-
-            //info!("Executing command with opcode {}", C::OPCODE);
-            let (slot, idx) = self
-                .slots
-                .acquire_slot(C::OPCODE, retval.as_mut_ptr())
-                .ok_or(CmdError::Param(param::Error::CONN_REJECTED_LIMITED_RESOURCES))?;
-            let _d = OnDrop::new(|| {
-                self.slots.release_slot(idx);
-            });
-            {
-                let mut io = self.writer.lock().await;
-                WithIndicator::new(cmd)
-                    .write_hci_async(io.deref_mut())
-                    .await
-                    .map_err(|e| CmdError::Controller(e.kind()))?;
-            }
-            let result = slot.wait().await;
-            let return_param_bytes = RemainingBytes::from_hci_bytes_complete(&retval[..result.len])
-                .map_err(|_| CmdError::Controller(embedded_io::ErrorKind::Other))?;
-            let e = CommandComplete {
-                num_hci_cmd_pkts: result.num_hci_cmd_pkts,
-                status: result.status,
-                cmd_opcode: C::OPCODE,
-                return_param_bytes,
-            };
-            let r = e.to_result::<C>().map_err(CmdError::Param)?;
-            // info!("Done executing command with opcode {}", C::OPCODE);
-            Ok(r)
-        }
-    }
-}
-
-impl<M, R, W, C, const SLOTS: usize> ControllerCmdAsync<C> for SerialController<M, R, W, SLOTS>
-where
-    M: RawMutex,
-    R: embedded_io_async::Read,
-    W: embedded_io_async::Write,
-    C: cmd::AsyncCmd,
-{
-    fn exec(&self, cmd: &C) -> impl Future<Output = Result<(), CmdError<Self::Error>>> {
-        async {
-            let mut io = self.writer.lock().await;
-            WithIndicator::new(cmd)
-                .write_hci_async(io.deref_mut())
+            let mut w = self.writer.lock().await;
+            w.write_all(tx)
                 .await
-                .map_err(|e| CmdError::Controller(e.kind()))?;
+                .map_err(|e| Error::Write(WriteHciError::Write(e)))?;
             Ok(())
         }
-    }
-}
-
-use core::mem;
-use core::mem::MaybeUninit;
-
-/// A type to delay the drop handler invocation.
-#[must_use = "to delay the drop handler invocation to the end of the scope"]
-pub struct OnDrop<F: FnOnce()> {
-    f: MaybeUninit<F>,
-}
-
-impl<F: FnOnce()> OnDrop<F> {
-    /// Create a new instance.
-    pub fn new(f: F) -> Self {
-        Self { f: MaybeUninit::new(f) }
-    }
-
-    /// Prevent drop handler from running.
-    pub fn defuse(self) {
-        mem::forget(self)
-    }
-}
-
-impl<F: FnOnce()> Drop for OnDrop<F> {
-    fn drop(&mut self) {
-        unsafe { self.f.as_ptr().read()() }
-    }
-}
-
-/// An explosive ordinance that panics if it is improperly disposed of.
-///
-/// This is to forbid dropping futures, when there is absolutely no other choice.
-///
-/// To correctly dispose of this device, call the [defuse](struct.DropBomb.html#method.defuse)
-/// method before this object is dropped.
-#[must_use = "to delay the drop bomb invokation to the end of the scope"]
-pub struct DropBomb {
-    _private: (),
-}
-
-impl DropBomb {
-    /// Create a new instance.
-    pub fn new() -> Self {
-        Self { _private: () }
-    }
-
-    /// Defuses the bomb, rendering it safe to drop.
-    pub fn defuse(self) {
-        mem::forget(self)
-    }
-}
-
-impl Drop for DropBomb {
-    fn drop(&mut self) {
-        panic!("boom")
     }
 }
